+++ multiqc/modules/pairtools/pairtools.py
@@ -1,0 +1,41 @@
+#!/usr/bin/env python
+
+""" MultiQC module to parse stats output from pairtools """
+
+import logging
+
+from multiqc.modules.base_module import BaseMultiqcModule
+
+# Initialise the logger
+log = logging.getLogger(__name__)
+
+class MultiqcModule(BaseMultiqcModule):
+    """This MultiQC module parses various
+    stats produced by pairtools."""
+
+    def __init__(self):
+
+        # Initialise the parent object
+        super(MultiqcModule, self).__init__(name='pairtools', anchor='pairtools',
+        href="https://github.com/mirnylab/pairtools",
+        info="pairtools is a command-line framework to process sequencing data from a Hi-C experiment.")
+
+
+        # Find and load any pairtools stats summaries
+        self.pairtools_stats = dict()
+            for f in self.find_log_files('pairtools'):
+                self.parse_pairtools_stats(f)
+
+        # Filter to strip out ignored sample names
+        self.pairtools_stats = self.ignore_samples(self.pairtools_stats)
+
+        if len(self.pairtools_stats) == 0:
+            raise UserWarning
+
+        log.info("Found {} reports".format(len(self.pairtools_stats)))
+
+
+    def parse_hicup_logs(self, f):
+        """ Parse a pairtools summary stats """
+        lof.info("parsing {} ...".format(f))
+        pass